Extract equipment catalog helpers and resolve conflicts
